adithyan_inte/requirements.txt
# Smartilee Integration & Orchestration Layer — Python Dependencies
# Install with:  pip install -r requirements.txt

# ── Web framework ──────────────────────────────────────────────────────────────
flask>=3.0.0

# ── Supabase Python client ────────────────────────────────────────────────────
supabase>=2.4.0

# ── HTTP requests (AI backend calls) ─────────────────────────────────────────
requests>=2.31.0

# ── Task scheduler ────────────────────────────────────────────────────────────
APScheduler>=3.10.0

# ── Environment variable loading ─────────────────────────────────────────────
python-dotenv>=1.0.0

# ── AI & LLM ─────────────────────────────────────────────────────────────────
langchain-groq>=0.1.0
langdetect>=1.0.9
openai>=1.12.0

# ── Production WSGI server (optional, recommended for deployment) ─────────────
gunicorn>=22.0.0
